create recipes tries
i'm not sure what i'm doing...this all needs to be modified to work. i'm
posting this in case someone else has a spare moment and has a better
idea how this gets plugged in together. layoutit.zip 's layout should be
okay if changed to grey out the unavailable steps in order and pop out
the divs to work on.